AI Development/Projects/Solutions/Module1-WorkingWithLocalModels/Exercise2-ParameterPlayground/requirements.txt
gradio==5.29.0
ollama==0.6.0
llama-cpp-python==0.3.4
